added min_target_points to base processor

+++ src/models_architecture/base_xgb_model.py
@@ -64,8 +64,8 @@ class XGBTrainModel(BaseModel):
             num_eval_batches = -1
         for batch_x, batch_y in eval_ds.take(num_eval_batches):
             X = np.stack([tf.squeeze(value) for value in batch_x.values()], axis=-1)
-            y_true = tf.squeeze((batch_y)) #np.argmax(batch_y, axis=1)
-            y_pred = tf.squeeze(self.xgb_model.predict(X))
+            y_true = tf.cast(tf.squeeze((batch_y)), tf.int32) #np.argmax(batch_y, axis=1)
+            y_pred = tf.squeeze(tf.cast(self.xgb_model.predict(X), tf.int32))
             y_pred = tf.one_hot(y_pred, depth=3)
 
             precision_buy.update_state(y_true, y_pred)
